Combat system - abilities
Implemented the actual combat system movement. Albeit a bit janky, the user can move a combatant, that combatant than has to confirm or cancel. The player then selects an enemy and then the ability they are going to use. After this it moves to the next combatant and repeats the process. TODOs - change the behaviour for enemies, add enum state, add better UI, add ability range. (Might get ability area added in)