Initiative tracker
Created the initiative UI. It works for the most part except the transition from the hedgehog to the main character does not happen. To use initiative attach the script to an image and the battle manager. (TO DO: Get rid of nasty public variables and create combatant outline when hovering over UI initiative icon